refactor: extract xtdb delta
